refactor: share chain id normalization helpers
